requirements.txt
Flask
requests
numpy
faiss-cpu
huggingface_hub
pywebview
